Made fixes recommended in Pull request
