+++ timApp/tim.py
@@ -77,6 +77,7 @@ from timApp.scheduling.scheduling_routes import scheduling
 from timApp.securitytxt.routes import securitytxt
 from timApp.sisu.scim import scim
 from timApp.sisu.sisu import sisu
+from timApp.tide_support.routes import ide
 from timApp.tim_app import app
 from timApp.timdb.sqa import db
 from timApp.upload.upload import upload
@@ -160,6 +161,7 @@ blueprints = [
     verify,
     quantum_circuit_plugin,
     symbolbutton_plugin,
+    ide,
 ]
 
 if app.config["BOOKMARKS_ENABLED"]:
@@ -286,8 +288,8 @@ def install_sql_hook():
             prev_exec_time = curr
             for r in traceback.format_stack():
                 if (
-                    r.startswith('  File "/service/')
-                    and not receive_before_execute.__name__ in r
+                        r.startswith('  File "/service/')
+                        and not receive_before_execute.__name__ in r
                 ):
                     print(r, end="")
             try:
@@ -333,9 +335,9 @@ def preprocess_request():
                 referrer_domain = None
             last_referrers = session.get("last_referrers", [])
             if (
-                referrer_domain
-                and referrer_domain != request.host
-                and referrer_domain not in last_referrers
+                    referrer_domain
+                    and referrer_domain != request.host
+                    and referrer_domain not in last_referrers
             ):
                 last_referrers.append(referrer_domain)
                 # Only leave the last 5 referrers.
